Use fixed-width types and fix radio/EEPROM handling
Replace generic Arduino types (int, unsigned int, short, byte, unsigned char, unsigned long) with fixed-width types (uint8_t, uint16_t, int8_t, uint32_t) across the project for better type safety and predictable memory usage. Add include guard and stdint.h to Config.h, convert rc_packet to uint16_t[], correct telemetry struct name/fields, and update radio read/write to use sizeof(rc_packet)/sizeof(telemetry_packet) for correct payload sizes. Convert loop indices and local vars to uint8_t/uint16_t as appropriate, adjust map/temporary variable types, and tidy a few initializations (modelNameSelection placement, hex constant case). These changes improve portability, reduce implicit promotions, and fix potential bugs in radio communication and EEPROM indexing.

+++ TX_nRF24_2ch_OLED/Calibration.ino
@@ -8,7 +8,7 @@ void calibration()
   while (calibStatus == 1)
   {
     // Get min. and max. value for each channel
-    for (int i = 0; i < RC_CHANNELS; i++)
+    for (uint8_t i = 0; i < RC_CHANNELS; i++)
     {
       raw_pots = analogRead(pins_pots[i]);
       
@@ -47,7 +47,7 @@ void calibration()
   while (calibStatus == 1)
   {
     // Get center value for each channel
-    for (int i = 0; i < RC_CHANNELS; i++)
+    for (uint8_t i = 0; i < RC_CHANNELS; i++)
     {
       mid_pots_calib[i] = analogRead(pins_pots[i]);
     }
